Work in progress: Android SDK compatibility

+++ org.eclipse.jgit/src/org/eclipse/jgit/internal/storage/file/ObjectDirectory.java
@@ -611,16 +611,11 @@ public class ObjectDirectory extends FileObjectDatabase {
 			FileUtils.delete(tmp, FileUtils.RETRY);
 			return InsertLooseObjectResult.EXISTS_LOOSE;
 		}
-		try {
-			Files.move(tmp.toPath(), dst.toPath(),
-					StandardCopyOption.ATOMIC_MOVE);
+        boolean ok = tmp.renameTo(dst);
+		if (ok) {
 			dst.setReadOnly();
 			unpackedObjectCache.add(id);
 			return InsertLooseObjectResult.INSERTED;
-		} catch (AtomicMoveNotSupportedException e) {
-			LOG.error(e.getMessage(), e);
-		} catch (IOException e) {
-			// ignore
 		}
 
 		// Maybe the directory doesn't exist yet as the object
@@ -628,17 +623,12 @@ public class ObjectDirectory extends FileObjectDatabase {
 		// try the rename first as the directory likely does exist.
 		//
 		FileUtils.mkdir(dst.getParentFile(), true);
-		try {
-			Files.move(tmp.toPath(), dst.toPath(),
-					StandardCopyOption.ATOMIC_MOVE);
+        ok = tmp.renameTo(dst);
+        if (ok) {
 			dst.setReadOnly();
-			unpackedObjectCache.add(id);
-			return InsertLooseObjectResult.INSERTED;
-		} catch (AtomicMoveNotSupportedException e) {
-			LOG.error(e.getMessage(), e);
-		} catch (IOException e) {
-			LOG.debug(e.getMessage(), e);
-		}
+            unpackedObjectCache.add(id);
+            return InsertLooseObjectResult.INSERTED;
+        }
 
 		if (!createDuplicate && has(id)) {
 			FileUtils.delete(tmp, FileUtils.RETRY);
